#97: Move TransactionAttributeType and remove json value from it
