+++ core/src/main/java/io/neow3j/transaction/TransactionAttribute.java
@@ -4,7 +4,6 @@ import io.neow3j.io.BinaryReader;
 import io.neow3j.io.BinaryWriter;
 import io.neow3j.io.NeoSerializable;
 import io.neow3j.io.exceptions.DeserializationException;
-import io.neow3j.model.types.TransactionAttributeType;
 import java.io.IOException;
 import java.util.Objects;
